+++ project/src/pages/offer-screen/offer-screen.tsx
@@ -3,7 +3,7 @@ import { Helmet } from 'react-helmet-async';
 import { CommentForm } from '../../components/comment-form/comment-form';
 import ReviewsList from '../../components/reviews-list/reviews-list';
 import { ReviewType } from '../../mocks/reviews';
-import { Link, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { useState } from 'react';
 import {
   getOfferOnPage,
@@ -14,7 +14,7 @@ import Map from '../../components/map/map';
 import Card from '../../components/card/card';
 import { OfferType } from '../../mocks/offers';
 import { useAppSelector } from '../../hooks';
-import { AppRoutes, AuthorizationsStatus } from '../../consts';
+import HeaderNav from '../../components/header-nav/header-nav';
 
 type OfferScreenProps = {
   reviews: ReviewType[];
@@ -23,7 +23,6 @@ type OfferScreenProps = {
 export default function OfferScreen({
   reviews,
 }: OfferScreenProps): JSX.Element {
-  const status = useAppSelector((state) => state.authorizationStatus);
   const offersInThisCity = useAppSelector(
     (state) => state.offersForCurrentCity
   );
@@ -70,35 +69,7 @@ export default function OfferScreen({
             </div>
             <nav className="header__nav">
               <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  {status === AuthorizationsStatus.Auth ? (
-                    <>
-                      <Link
-                        className="header__nav-link header__nav-link--profile"
-                        to={AppRoutes.Favorites}
-                      >
-                        <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                        <span className="header__user-name user__name">
-                          {localStorage.getItem('user')}
-                        </span>
-                        <span className="header__favorite-count">3</span>
-                      </Link>
-                      <a className="header__nav-link" href="#">
-                        <span className="header__signout">Sign out</span>
-                      </a>
-                    </>
-                  ) : (
-                    <li className="header__nav-item">
-                      <Link
-                        className="header__nav-link header__nav-link--profile"
-                        to={AppRoutes.Login}
-                      >
-                        <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                        <span className="header__login">Sign in</span>
-                      </Link>
-                    </li>
-                  )}
-                </li>
+                <HeaderNav />
               </ul>
             </nav>
           </div>
