+++ project/src/types/state.ts
@@ -10,7 +10,6 @@ export type AuthData = {
 };
 
 export type UserData = {
-  id: number;
   email: string;
   token: string;
 };
